requirements.txt
pillow==11.1.0
colorama==0.4.6
qrcode==8.0
numpy==2.2.2
opencv-python==4.11.0.86